feature/global_ens_atmos_plots (#233)
* Use 'bs' for Brier Score

Refs: #5

* Update for NAEFS grid2grid plots to follow EVS naming conventions

Refs: #5

* Update for NAEFS grid2obs plots to follow EVS naming conventions

Refs: #5

* Update for NAEFS precip plots to follow EVS naming conventions

Refs: #5

* Remove logging settings

Refs: #5

* Update for global_ens grid2grid plots to follow EVS naming conventions

Refs: #5

* Check SENDCOM value

Refs: #5

* More updates for global_ens grid2grid plots to follow EVS naming conventions

Refs: #5

* Update for global_ens grid2obs init separate plots to follow EVS naming conventions

Refs: #5

* More updates for global_ens grid2obs separate init plots to follow EVS naming conventions

Refs: #5

* Updates for global_ens grid2obs plots to follow EVS naming conventions

Refs: #5

* More updates for global_ens precip plots to follow EVS naming conventions

Refs: #5

* Revamp renaming of NAEFS graphics

Refs: #5

* Redo image renaming in exevs_global_ens_atmos_gefs_grid2grid_plots.sh

Refs: #5

* Redo image renaming and minor fixes in exevs_global_ens_atmos_gefs_grid2obs_plots.sh

Refs: #5

* Redo image renaming and minor fixes in exevs_global_ens_atmos_gefs_grid2obs_separate_plots.sh

Refs: #5

* Redo image renaming in exevs_global_ens_atmos_gefs_precip_plots.sh

Refs: #5

* Updates for gefs precip spatial plots

Refs: #5

* Redo image renaming in exevs_global_ens_atmos_gefs_profile1_plots.sh

Refs: #5

* Remove UGRD/P400 and VGRD/P400 plotting in exevs_global_ens_atmos_gefs_profile1_plots.sh

Refs: #5

* Redo image renaming and remove UGRD/P400 and VGRD/P400 in exevs_global_ens_atmos_gefs_profile2_plots.sh

Refs: #5

* Redo image renaming and remove UGRD/P400 and VGRD/P400 in exevs_global_ens_atmos_gefs_profile3_plots.sh

Refs: #5

* Redo image renaming and remove UGRD/P400 and VGRD/P400 in exevs_global_ens_atmos_gefs_profile4_plots.sh

Refs: #5

* Updates for global_ens sea-ice plots to follow EVS naming conventions

Refs: #5

* Updates for global_ens sea-ice plots to follow EVS naming conventions

Refs: #5

* Redo image renaming in exevs_global_ens_atmos_gefs_sst_plots.sh

Refs: #5

* Rename driver script files

Refs: #5

* Fix formatting of valid hour in graphic name

Refs: #5

* Fix global_ens wave plots image names

Refs: #5

* Updates for the headline plots

Refs: #5

* Adjust global_ens wave plots resources

Refs: #5

* Adjust global_ens headline plots resources

Refs: #5

* Fix global_ens atmos NAEFS tar file names

Refs: #5

* Adjust global_ens atmos NAEFS plots job resources

Refs: #5

* Fix global_ens atmos GEFS sea-ice, snowfall, sst tar file names

Refs: #5

* Adjust global_ens atmos GEFS sea-ice, snowfall, sst plots job resources

Refs: #5

* Fix global_ens atmos GEFS profile tar file names and mpiexec commands

Refs: #5

* Adjust global_ens atmos GEFS profile plots job resources

Refs: #5

* Adjust global_ens atmos GEFS precip plots job resources

Refs: #5

* Fix global_ens atmos GEFS precip tar file names

Refs: #5

* Remove unused versions/ run files

* Adjust global_ens atmos GEFS grid2grid and grid2obs plots job resources

Refs: #5

* Fix global_ens atmos GEFS grid2grid and grid2obs tar file names and mpiexec commands

Refs: #5

* Remove unused j-jobs and ex-scripts

Refs: #5

* Add F000 to plots

Refs: #5

* Add Tropics display title name

Refs: #5

* Fix sea-ice concentration title

Refs: #5

* Fix sea-ice concentration title in perf_diag

Refs: #5

* Fix typo

Refs: #5

* Fix SST title

Refs: #5

* Remove unused files

Refs: #8

* Remove METplus load for global_ens plots

Refs: #5

* Remove default setting of NDATE to WCOSS1 path

Refs: #5

* Fix wave plot titles

Refs: #8

* Fix copy commands

Refs: #8

* Add back in load of METplus

Refs: #5

* Remove hard-coded Cartopy path

Refs: #5

* Remove unused script

Refs: #5

* Clean up jevs_global_ens_wave_grid2obs_plots.sh

Refs: #8

* Correct xticks precision

Refs: #5

* Fix image names

Refs: #5

* Fix units for upper-level TMP and PRMSL

Refs: #5

* Remove function to get get_d data

Refs: #5

+++ scripts/global_ens/plots/exevs_global_ens_atmos_gefs_precip_spatial_plots.sh
@@ -55,34 +55,25 @@ for model in $model_list ; do
   past=`$NDATE -$fhr ${VDATE}12`	
   INITDATE=${past:0:8}
   apcp24mean=$source/GenEnsProd_${MODEL}_APCP24_FHR${fhr}_${VDATE}_120000V_ens.nc
-  size=`ls -l $apcp24mean | awk '{ print $5}'`
-  if [ $size -gt 1000000 ] ; then
-    ln -sf $source/GenEnsProd_${MODEL}_APCP24_FHR${fhr}_${VDATE}_120000V_ens.nc $target/${model}_precip_24hrAccum_init${INITDATE}12_fhr${fhr}.nc	
+  if [ -f $apcp24mean ]; then
+      size=`ls -l $apcp24mean | awk '{ print $5}'`
+      if [ $size -gt 1000000 ] ; then
+          ln -sf $source/GenEnsProd_${MODEL}_APCP24_FHR${fhr}_${VDATE}_120000V_ens.nc $target/${model}_precip_24hrAccum_init${INITDATE}12_fhr${fhr}.nc
+      else
+          echo "$apcp24mean SIZE LESS THAN 1000000"
+      fi
+  else
+      echo "$apcp24mean DOES NOT EXIST"
   fi 
  done
 done
-
-source=$COMINccpa24/gefs
-target=$DATA/grid2grid_plots/data/ccpa
-mkdir -p $target
-if [ -s $source/ccpa.t12z.grid3.24h.f00.nc ] ; then
- ln -sf $source/ccpa.t12z.grid3.24h.f00.nc $target/ccpa_precip_24hrAccum_valid${VDATE}12.nc
-fi 
 
 python $USHevs/global_ens/ush_gens_plot_py/global_det_atmos_plots.py
 
 cd $DATA/grid2grid_plots/plot_output/atmos.${VDATE}/precip/SL1L2_FBAR_24hrAccumMaps_CONUS_precip_spatial_map/images
 
-#scp *.png wd20bz@emcrzdm:/home/people/emc/www/htdocs/bzhou/evs_plots/gens/spaticl
+tar -cvf evs.plots.${COMPONENT}.${RUN}.${MODELNAME}.precip_spatial.v${VDATE}.tar *.gif
 
-mv qpe.v${VDATE}12.024h.conus.png evs.ccpa.spatial_map.apcp_a24_vlid12z_f000.conus.png
-
-tar -cvf evs.plots.gefs.precip.spatial.map.v${VDATE}.tar *.gif
-
-cp evs.plots.gefs.precip.spatial.map.v${VDATE}.tar  $COMOUT/.  
-
-
-
-
-
-
+if [ $SENDCOM = YES ]; then
+    cp evs.plots.${COMPONENT}.${RUN}.${MODELNAME}.precip_spatial.v${VDATE}.tar  $COMOUT/.
+fi
